move RPN functions around to model_rpn; improve logging
