+++ examples/FasterRCNN/modeling/model_rpn.py
@@ -1,11 +1,13 @@
 # -*- coding: utf-8 -*-
 
 import tensorflow as tf
+import numpy as np
 
 from tensorpack.models import Conv2D, layer_register
 from tensorpack.tfutils.argscope import argscope
 from tensorpack.tfutils.scope_utils import auto_reuse_variable_scope, under_name_scope
 from tensorpack.tfutils.summary import add_moving_summary
+from tensorpack.utils.argtools import memoized
 
 from config import config as cfg
 from .model_box import clip_boxes
@@ -151,3 +153,47 @@ def generate_rpn_proposals(boxes, scores, img_shape,
     proposal_scores = tf.gather(topk_valid_scores, nms_indices)
     tf.sigmoid(proposal_scores, name='probs')  # for visualization
     return tf.stop_gradient(proposal_boxes, name='boxes'), tf.stop_gradient(proposal_scores, name='scores')
+
+
+@memoized
+def get_all_anchors(*, stride, sizes, ratios, max_size):
+    """
+    Get all anchors in the largest possible image, shifted, floatbox
+    Args:
+        stride (int): the stride of anchors.
+        sizes (tuple[int]): the sizes (sqrt area) of anchors
+        ratios (tuple[int]): the aspect ratios of anchors
+        max_size (int): maximum size of input image
+
+    Returns:
+        anchors: SxSxNUM_ANCHORx4, where S == ceil(MAX_SIZE/STRIDE), floatbox
+        The layout in the NUM_ANCHOR dim is NUM_RATIO x NUM_SIZE.
+
+    """
+    # Generates a NAx4 matrix of anchor boxes in (x1, y1, x2, y2) format. Anchors
+    # are centered on 0, have sqrt areas equal to the specified sizes, and aspect ratios as given.
+    anchors = []
+    for sz in sizes:
+        for ratio in ratios:
+            w = np.sqrt(sz * sz / ratio)
+            h = ratio * w
+            anchors.append([-w, -h, w, h])
+    cell_anchors = np.asarray(anchors) * 0.5
+
+    field_size = int(np.ceil(max_size / stride))
+    shifts = (np.arange(0, field_size) * stride).astype("float32")
+    shift_x, shift_y = np.meshgrid(shifts, shifts)
+    shift_x = shift_x.flatten()
+    shift_y = shift_y.flatten()
+    shifts = np.vstack((shift_x, shift_y, shift_x, shift_y)).transpose()
+    # Kx4, K = field_size * field_size
+    K = shifts.shape[0]
+
+    A = cell_anchors.shape[0]
+    field_of_anchors = cell_anchors.reshape((1, A, 4)) + shifts.reshape((1, K, 4)).transpose((1, 0, 2))
+    field_of_anchors = field_of_anchors.reshape((field_size, field_size, A, 4))
+    # FSxFSxAx4
+    # Many rounding happens inside the anchor code anyway
+    # assert np.all(field_of_anchors == field_of_anchors.astype('int32'))
+    field_of_anchors = field_of_anchors.astype("float32")
+    return field_of_anchors
